refactor: move builder logic to seperate file
